feat: update avatar image if needed and extract components
